feat: refactor api to have nested routing and less logic
