+++ src/routes.js
@@ -1,13 +1,8 @@
 import usersRouter from './users';
-import workoutsRouter from './workouts';
 
 export default [
   {
     path: '/users',
     handler: usersRouter,
-  },
-  {
-    path: '/workouts',
-    handler: workoutsRouter,
   },
 ];
