+++ src/users/controller/update.js
@@ -1,11 +1,20 @@
 import User from '../model';
 
+import {
+  isValidObjectId,
+} from '../../utilities';
+import {
+  invalidIdError,
+} from '../../utilities/errors';
+
 export default function update(req, res, next) {
-  const { id } = req.params;
+  const { userId } = req.params;
   const updates = req.body;
 
-  User
-    .findByIdAndUpdate(id, updates, { new: true })
+  if (!isValidObjectId(userId)) return next(invalidIdError);
+
+  return User
+    .findByIdAndUpdate(userId, updates, { new: true })
     .exec((err, user) => {
       if (err) return next(err);
