+++ src/workouts/controller/update.js
@@ -1,11 +1,19 @@
 import Workout from '../model';
+import {
+  isValidObjectId,
+} from '../../utilities';
+import {
+  invalidIdError,
+} from '../../utilities/errors';
 
 export default function update(req, res, next) {
-  const { id } = req.params;
+  const { workoutId } = req.params;
   const updates = req.body;
 
-  Workout
-    .findByIdAndUpdate(id, updates, { new: true })
+  if (!isValidObjectId(workoutId)) return next(invalidIdError);
+
+  return Workout
+    .findByIdAndUpdate(workoutId, updates, { new: true })
     .exec((err, workout) => {
       if (err) return next(err);
